consolidate newsletter utils
